chore(): Move components into page files
The auth pages had "panel" components that were not reusable.
This commit pulls the contents of the panel components into
the pages files for easier reading and less fragmentation.
