+++ pages/login.tsx
@@ -1,13 +1,40 @@
-import { NextPage } from "next";
 import Head from "next/head";
+import Link from "next/link";
+import { Formik, FormikHelpers, Form } from "formik";
+import { NextPage } from "next";
+import { useRouter } from "next/router";
 
-import { LoginPanel } from 'auth'
+import { ILoginUser } from "auth/auth.types";
+import { IRequestError } from "common/types";
+import { Input } from "common/forms";
+import { userCurrentUser } from "auth";
 
 const LoginPage: NextPage = () => {
+  const router = useRouter();
+  const { login } = userCurrentUser();
+  const { redirect } = router.query;
+
+  const handleSubmit = async (
+    credentials: ILoginUser,
+    { setErrors, setFieldError }: FormikHelpers<ILoginUser>
+  ) => {
+    try {
+      await login(credentials);
+      router.push(typeof redirect === "string" ? redirect : "/");
+    } catch (error: any) {
+      const { message, fieldErrors }: IRequestError = error;
+      if (fieldErrors) {
+        setErrors(fieldErrors);
+      } else {
+        setFieldError("username", message);
+      }
+    }
+  };
+
   return (
     <>
       <Head>
-        <title>LoginPage the Gastro Cooperative</title>
+        <title>Gastro: Login</title>
         <meta
           name="description"
           content="A living history of the world's food written by its people"
@@ -15,7 +42,32 @@ const LoginPage: NextPage = () => {
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      <LoginPanel />
+      <Formik
+        initialValues={{ username: "", password: "" }}
+        onSubmit={handleSubmit}
+      >
+        <Form>
+          <h2>Login</h2>
+          <Input
+            name="username"
+            label="Username"
+            placeholder="Username / email"
+          />
+          <Input
+            type="password"
+            label="Password"
+            name="password"
+            placeholder="Password"
+          />
+          <button type="submit">Login</button>
+        </Form>
+      </Formik>
+      <Link href="/signup">
+        <a>Create an Account</a>
+      </Link>
+      <Link href="/reset-password">
+        <a>Forgot your password?</a>
+      </Link>
     </>
   );
 };
