+++ src/auth/index.ts
@@ -1,4 +1,1 @@
 export { userCurrentUser } from './userContextProvider'
-export { default as LoginPanel } from './loginPanel'
-export { default as SignupPanel } from './signupPanel'
-export { default as ResetPasswordPanel } from './resetPasswordPanel'
